Using size_t instead of unsigned long and uint8_t instead of char for 1 byte.

+++ src/battery.h
@@ -1,4 +1,6 @@
+#include <stddef.h>
+
 #pragma once
 
-void update_battery(unsigned long);
+void update_battery(size_t);
 void show_battery();
