+++ src/date_time.h
@@ -1,4 +1,6 @@
+#include <stddef.h>
+
 #pragma once
 
-void update_date_time(unsigned long);
+void update_date_time(size_t);
 void show_date_time();
