+++ src/utils.h
@@ -1,3 +1,5 @@
+#include <stddef.h>
+
 #pragma once
 
-void format_size(unsigned long);
+void format_size(size_t);
